+++ app/domain/sys/code/dao/code_sequence_dao.go
@@ -34,7 +34,7 @@ func NewCodeSequenceDao(dbKey string) *CodeSequenceDao {
 
 // NextSeq 获取下一个序列号
 // 如果当天记录不存在，会自动创建并返回 1
-func (d *CodeSequenceDao) NextSeq(ctx context.Context, typeCode, dateStr string, count int) (int64, error) {
+func (d *CodeSequenceDao) NextSeq(ctx context.Context, typeCode, dateStr string, count int64) (int64, error) {
 	if count <= 0 {
 		count = 1
 	}
@@ -48,7 +48,7 @@ func (d *CodeSequenceDao) NextSeq(ctx context.Context, typeCode, dateStr string,
 
 // NextSeq 获取下一个序列号
 // 如果当天记录不存在，会自动创建并返回 1
-func (d *CodeSequenceDao) mongoNextSeq(ctx context.Context, typeCode, dateStr string, count int) (int64, error) {
+func (d *CodeSequenceDao) mongoNextSeq(ctx context.Context, typeCode, dateStr string, count int64) (int64, error) {
 	mDao, ok := d.Dao.GetStore().(store_mongodb.IMongoDao[*model.CodeSequence])
 	if !ok {
 		return 0, fmt.Errorf("[code_sequence.dao] dao does not implement IMongoDao")
